[Design 🎨] 멤버들 간격 수정
